pyproject.toml
[project]
name = "cuda-jit-ptx-test"
version = "0.1.0"
description = "A Python package with CUDA-dependent functionality"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Kernel Machines", email = "kernel_machines@pm.me"}
]
dependencies = [
    "numpy>=1.21",
]

[project.optional-dependencies]
cu11 = [
    "cuda-core[cu11]>=0.1",
]
cu12 = [
    "cuda-core[cu12]>=0.1",
]

[project.urls]
Homepage = "https://github.com/Kernel-Machines/cuda-jit-ptx-test"
Repository = "https://github.com/Kernel-Machines/cuda-jit-ptx-test.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cuda_jit_ptx_test"]